:sparkles: Add `accountIndex` attribute to `LedgerDevice`
Allows a Ledger device to be initialized to sign transactions
for a particular account instead of always signing with the
default account for its given path.

Technically, this could have been achieved before by initializing
the device with a path wherein the last path element is offset by
the index of the desired account. This is cumbersome, however, and
adds some complexity (i.e: account path derivation) that does fit
well within the scope of the device's responsibilities.